Add files via upload
These are all of the brainstorming and presentation elements of our project! It was important that our project be well-organized and that all information be sourced accurately, so we documented all of our ideas and research with these documents. We also made sure that the technical aspects of this project were touched upon during the presentation before diving into how we could maintain and expand this project long-term.